requirements.txt
streamlit
google-generativeai
apify-client
pandas
python-dotenv